Moved `SpellLevel` into enums
