+++ gameserver/src/main/java/com/lhf/game/creature/vocation/DMV.java
@@ -7,10 +7,10 @@ import com.lhf.game.dice.DiceD20;
 import com.lhf.game.dice.MultiRollResult;
 import com.lhf.game.enums.Attributes;
 import com.lhf.game.enums.EquipmentTypes;
+import com.lhf.game.enums.SpellLevel;
 import com.lhf.game.enums.Stats;
 import com.lhf.game.item.concrete.HealPotion;
 import com.lhf.game.magic.CubeHolder;
-import com.lhf.game.magic.SpellLevel;
 
 public class DMV extends Vocation implements CubeHolder {
