+++ gameserver/src/main/java/com/lhf/game/creature/vocation/Mage.java
@@ -10,11 +10,11 @@ import com.lhf.game.dice.DiceD20;
 import com.lhf.game.dice.MultiRollResult;
 import com.lhf.game.enums.Attributes;
 import com.lhf.game.enums.EquipmentTypes;
+import com.lhf.game.enums.SpellLevel;
 import com.lhf.game.enums.Stats;
 import com.lhf.game.item.concrete.HealPotion;
 import com.lhf.game.item.concrete.equipment.LeatherArmor;
 import com.lhf.game.magic.CubeHolder;
-import com.lhf.game.magic.SpellLevel;
 
 public class Mage extends Vocation implements CubeHolder {
     private static class SpellSlot {
