+++ gameserver/src/main/java/com/lhf/game/magic/builder/CLIAdapter.java
@@ -18,8 +18,8 @@ import com.lhf.game.dice.DamageDice;
 import com.lhf.game.dice.DieType;
 import com.lhf.game.enums.Attributes;
 import com.lhf.game.enums.DamageFlavor;
+import com.lhf.game.enums.SpellLevel;
 import com.lhf.game.enums.Stats;
-import com.lhf.game.magic.SpellLevel;
 import com.lhf.game.magic.CreatureAOESpellEntry.AutoTargeted;
 import com.lhf.game.magic.builder.SpellEntryBuilder.SpellEntryBuilderAdapter;
