+++ gameserver/src/main/java/com/lhf/game/magic/concrete/ShockBolt.java
@@ -14,9 +14,9 @@ import com.lhf.game.dice.DamageDice;
 import com.lhf.game.dice.DieType;
 import com.lhf.game.enums.Attributes;
 import com.lhf.game.enums.DamageFlavor;
+import com.lhf.game.enums.SpellLevel;
 import com.lhf.game.enums.Stats;
 import com.lhf.game.magic.CreatureTargetingSpellEntry;
-import com.lhf.game.magic.SpellLevel;
 import com.lhf.messages.out.CastingMessage;
 
 public class ShockBolt extends CreatureTargetingSpellEntry {
