+++ gameserver/src/main/java/com/lhf/game/magic/concrete/ThunderStrike.java
@@ -15,9 +15,9 @@ import com.lhf.game.dice.DamageDice;
 import com.lhf.game.dice.DieType;
 import com.lhf.game.enums.Attributes;
 import com.lhf.game.enums.DamageFlavor;
+import com.lhf.game.enums.SpellLevel;
 import com.lhf.game.enums.Stats;
 import com.lhf.game.magic.CreatureTargetingSpellEntry;
-import com.lhf.game.magic.SpellLevel;
 import com.lhf.messages.out.CastingMessage;
 
 public class ThunderStrike extends CreatureTargetingSpellEntry {
